Add health to top nav
This is an implementation without adding a submenu or dropdown for
projects.

The only style change needed to set a minimum width on the name in the
top left corner, so that the longer top nav doesn't squish it as the
screen narrows.

This causes the mobile version of the menu to appear at a wider
breakpoint than it would have in the past.

There's also a side-effect of compressing the animation horizontally
that is still a mystery at the moment.